sum series
sum a series of numbers from King 102. At this point in my learning, I am reading a King program once and trying to write it. I forgot to initialize sum = 0 the first time. It took me a moment to understand why there is a second scanf call to address of n. It looks like that is necessary to read the values into sum and print sum. Otherwise, the values would not be summed.