13 prints the range of files
My super goal was to write a probram that takes command line arguments of range and filename, then prints back the contents of those lines from the file.

This exercise was the entrance criteria for systems programming class at harvard extension.

It was hard because I have only learned C self-taught from used textbooks bought online. Today is the first day of the class. Looking at the program 13 here, I meet the criteria. This is a big achievement.

The program is not perfect, but it meets the criteria. Score!